fw: fix keyboard interfering with VGA registers
cc65 generates (zp), Y addressing mode, which reads
some random memory during execution. It's ok for
memory access, but disastrous for I/O. It caused VGA
registers to be modified (as some of them do not handle
reads and access is always treated as write).

Rewrote keyboard I/O access to asm to fix the issue.